Use startup script in instance creation
