+++ plugin.yaml
@@ -10,7 +10,7 @@ plugins:
     # URL to archive containing the plugin or name of directory containing
     # the plugin if it is included in the the blueprint directory under the
     # "plugins" directory.
-    source: https://github.com/szpotona/cloudify-gcp-plugin/archive/add_security_groups.zip
+    source: https://github.com/szpotona/cloudify-gcp-plugin/archive/use_resources.zip
 
 node_types:
   cloudify.gcp.nodes.Instance:
