+++ plugin/gcp/instance.py
@@ -18,143 +18,6 @@ from plugin.gcp.service import GCPError
 from plugin.gcp.service import blocking
 
 
-class FirewallRule(GoogleCloudPlatform):
-    def __init__(self,
-                 config,
-                 logger,
-                 firewall,
-                 network):
-        """
-
-        :param auth:
-        :param project:
-        :param logger:
-        :param firewall: firewall dictionary with a following structure:
-        firewall = {'name': 'firewallname',
-                    'allowed: [{ 'IPProtocol': 'tcp', 'ports': ['80']}],
-                    'sourceRanges':['0.0.0.0/0'],
-                    'sourceTags':['tag'], (optional)
-                    'targetTags':['tag2'] (optional)
-                    }
-        ref. https://cloud.google.com/compute/docs/reference/latest/firewalls
-        :param network: network name the firewall rule is connected to
-        :return:
-        """
-        super(FirewallRule, self).__init__(config, logger)
-        self.firewall = firewall
-        self.network = network
-        self.name = self.firewall['name']
-
-    @blocking(True)
-    def create(self):
-        """
-        Create GCP firewall rule in a GCP network.
-        Global operation.
-
-        :return: REST response with operation responsible for the firewall rule
-        creation process and its status
-        """
-        self.logger.info('Create firewall rule')
-        self.firewall['network'] = 'global/networks/{0}'.format(self.network)
-        return self.compute.firewalls().insert(
-            project=self.project,
-            body=self.firewall).execute()
-
-    @blocking(True)
-    def delete(self):
-        """
-        Delete GCP firewall rule from GCP network.
-        Global operation.
-
-        :return: REST response with operation responsible for the firewall rule
-        deletion process and its status
-        """
-        self.logger.info('Delete firewall rule')
-        return self.compute.firewalls().delete(
-            project=self.project,
-            firewall=self.firewall['name']).execute()
-
-    @blocking(True)
-    def update(self):
-        self.logger.info('Update firewall rule')
-        return self.compute.firewalls().update(
-            project=self.project,
-            firewall=self.firewall['name'],
-            body=self.firewall).execute()
-
-    def list(self):
-        """
-        List GCP firewall rules in all networks.
-
-        :return: REST response with list of firewall rules in a project
-        """
-        self.logger.info('List firewall rules in project')
-        return self.compute.firewalls().list(
-            project=self.project).execute()
-
-    def wait_for_operation(self, operation, global_operation=True):
-        super(FirewallRule, self).wait_for_operation(operation,
-                                                     global_operation)
-
-
-class Network(GoogleCloudPlatform):
-    def __init__(self,
-                 config,
-                 logger,
-                 network):
-        super(Network, self).__init__(config, logger)
-        self.network = network
-        self.name = utils.get_gcp_resource_name(network['name'])
-
-    @blocking(True)
-    def create(self):
-        """
-        Create GCP network.
-        Global operation.
-
-        :return: REST response with operation responsible for the network
-        creation process and its status
-        """
-        self.logger.info('Create network')
-        return self.compute.networks().insert(project=self.project,
-                                              body=self.to_dict()).execute()
-
-    @blocking(True)
-    def delete(self):
-        """
-        Delete GCP network.
-        Global operation
-
-        :param network: network name
-        :return: REST response with operation responsible for the network
-        deletion process and its status
-        """
-        self.logger.info('Delete network')
-        return self.compute.networks().delete(
-            project=self.project,
-            network=self.name).execute()
-
-    def list(self):
-        """
-        List networks.
-
-        :return: REST response with list of networks in a project
-        """
-        self.logger.info('List networks')
-        return self.compute.networks().list(
-            project=self.project).execute()
-
-    def wait_for_operation(self, operation, global_operation=True):
-        super(Network, self).wait_for_operation(operation, global_operation)
-
-    def to_dict(self):
-        body = {
-            'description': 'Cloudify generated network',
-            'name': self.name
-        }
-        return body
-
-
 class Instance(GoogleCloudPlatform):
     def __init__(self,
                  config,
@@ -182,7 +45,7 @@ class Instance(GoogleCloudPlatform):
         Create GCP VM instance with given parameters.
         Zone operation.
 
-        :param startup_script: shell script to be run on instance startup,
+        :param startup_script: shell script text to be run on instance startup,
         default None
         :return: REST response with operation responsible for the instance
         creation process and its status
